Move all graphs to origin tab
