+++ public/js/components/Tabs/Tabs.js
